requirements-dev.txt
pytest>=7.4.0
pytest-spark>=0.6.0
pyspark>=3.4.0
delta-spark>=2.4.0
chispa>=0.9.0
coverage>=7.0.0
pytest-cov>=4.1.0
